config/requirements.txt
aiogram>=3.0.0
python-dotenv>=1.0.0
openpyxl>=3.1.0
pytest>=7.0.0
pytest-asyncio>=0.21.0
